Fixed kill method.
Kill request first. Then wait and force it. Latter doesn't remove the tray icon, but sometimes is necessary.

+++ ESOLauncherKiller.py
@@ -10,6 +10,16 @@ while True:
 
     if foundGame and foundLauncher:
         # k = os.popen('taskkill /IM Bethesda.net_Launcher.exe').read()
-        os.popen('taskkill /F /IM Bethesda.net_Launcher.exe')
+        os.popen('taskkill /IM Bethesda.net_Launcher.exe')
+
+        # Verify kill!
+        sleep(10)
+
+        foundLauncher = None
+        foundLauncher = os.popen('tasklist /FI "IMAGENAME eq Bethesda.net_Launcher.exe" 2>NUL | \
+        find /I /N "Bethesda.net_Launcher.exe"').read()
+
+        if foundLauncher:
+            os.popen('taskkill /F /IM Bethesda.net_Launcher.exe')
 
     sleep(60 * 1)  # Minutes.
